+++ SherlDog/Extension+/ResizedImage+.swift
@@ -1,0 +1,17 @@
+//
+//  ResizedImage+.swift
+//  SherlDog
+//
+//  Created by JIN LEE on 7/18/25.
+//
+
+import UIKit
+
+extension UIImage {
+    func resized(to size: CGSize) -> UIImage {
+        let renderer = UIGraphicsImageRenderer(size: size)
+        return renderer.image { _ in
+            self.draw(in: CGRect(origin: .zero, size: size))
+        }
+    }
+}
