+++ SherlDog/Reusable/Component/BottomTabBarController.swift
@@ -65,12 +65,3 @@ class BottomTabBarController: UITabBarController {
         UITabBar.appearance().tintColor = .keycolorInverse
     }
 }
-
-extension UIImage {
-    func resized(to size: CGSize) -> UIImage {
-        let renderer = UIGraphicsImageRenderer(size: size)
-        return renderer.image { _ in
-            self.draw(in: CGRect(origin: .zero, size: size))
-        }
-    }
-}
